abstracting test ops/traits into helpers
