+++ fusion/src/Fusion/tests/autodiff/test_graph.cpp
@@ -1,5 +1,7 @@
 #include <gtest/gtest.h>
 
+#include "helpers.h"
+
 #include "Fusion/autodiff/Graph.hpp"
 
 template <typename T>
@@ -48,108 +50,6 @@ public:
         return graph.edges_;
     }
 };
-
-/// the below code is to be moved to fixture files on creation of opschema/op layer
-/// test suite
-
-struct UnaryTag{};
-struct BinaryTag{};
-struct SplitTag{};
-
-template <> struct OpTraits<UnaryTag> {
-   static constexpr std::string_view name = "TestUnaryOp";
-   static constexpr OpSchema schema{
-      .category = OpCategory::EwiseUnary,
-      .inputs = {.kind=ArityKind::Fixed, .arity=1},
-      .outputs = {.kind=ArityKind::Fixed, .arity=1},
-      .mutation = MutationKind::OutOfPlace,
-  };
-};
-
-
-template <> struct OpTraits<BinaryTag> {
-   static constexpr std::string_view name = "TestBinaryOp";
-   static constexpr OpSchema schema{
-      .category = OpCategory::EwiseBinary,
-      .inputs = {.kind=ArityKind::Fixed, .arity=2},
-      .outputs = {.kind=ArityKind::Fixed, .arity=1},
-      .mutation = MutationKind::OutOfPlace,
-  };
-};
-
-
-template <> struct OpTraits<SplitTag> {
-   static constexpr std::string_view name = "TestSplitOp";
-   static constexpr OpSchema schema{
-      .category = OpCategory::EwiseBinary,
-      .inputs = {.kind=ArityKind::Fixed, .arity=2},
-      .outputs = {.kind=ArityKind::Fixed, .arity=2},
-      .mutation = MutationKind::OutOfPlace,
-  };
-};
-
-
-
-template <typename T>
-struct TestUnaryOp {
-      using tag = UnaryTag;
-      using In = AutodiffMeta<T>;
-      using Out = AutodiffMeta<T>;
-      using GradIn = AutodiffMeta<T>;
-      using GradOut = AutodiffMeta<T>;
-
-   Out forward(Context<T> &context, const In &input) {
-      Out out;
-      return out;
-   }
-
-   GradIn backward(Context<T> &context, GradOut &grad_out) {
-      GradIn g;
-      return g;
-   }
-};
-
-template <typename T>
-struct TestBinaryOp {
-    using tag = BinaryTag;
-    using In = AutodiffMeta<T>;
-    using Out = AutodiffMeta<T>;
-    using GradIn = AutodiffMeta<T>;
-    using GradOut = AutodiffMeta<T>;
-
-   Out forward(Context<T> &context, const In &input) {
-       Out out;
-       return out;
-    }
-
-   GradIn backward(Context<T> &context, GradOut &grad_out) {
-       GradIn g;
-       return g;
-    }
-
-};
-
-template <typename T>
-struct TestSplitOp {
-   using tag = SplitTag;
-   using In = AutodiffMeta<T>;
-   using Out = AutodiffMeta<T>;
-   using GradIn = AutodiffMeta<T>;
-   using GradOut = AutodiffMeta<T>;
-
-   Out forward(Context<T> &context, const In &input) {
-      Out out;
-      return out;
-   }
-
-   GradIn backward(Context<T> &context, GradOut &grad_out) {
-      GradIn g;
-      return g;
-   }
-};
-
-
-
 
 class GraphTest : public ::testing::Test {
 protected:
